fix(vito2mqtt): type plant_power_output as float and tighten handler factory return
PR #232 review follow-ups:

- plant_power_output: the PR3 codec is (first byte / 2), so an odd raw
  byte decodes to a half-percent value. The int field rejected that
  under state_model validation (cosalette 0.9.0 ADR-068), publishing
  ReturnValidationError instead of state. Typed float now; schema
  property type integer -> number. A full audit of the seven group
  models found no other int/float codec mismatch. Closes cap-v0i.
- make_telemetry_handler now returns Callable[[OptolinkPort], Awaitable[Any]]
  so callers are type-checked against the closure's real signature
  (Copilot review comment).

+++ apps/vito2mqtt/packages/src/vito2mqtt/devices/telemetry_models.py
@@ -193,8 +193,11 @@ class BurnerState:
             )
         ),
     ]
+    # float, not int: the PR3 codec is (first byte / 2), so an odd raw byte
+    # decodes to a half-percent value (e.g. 20.5). An int field would reject
+    # that under state_model validation and publish ReturnValidationError.
     plant_power_output: Annotated[
-        int, Field(json_schema_extra=_percent("Plant Power Output", icon="mdi:gauge"))
+        float, Field(json_schema_extra=_percent("Plant Power Output", icon="mdi:gauge"))
     ]
